refactor(knowledge): consolidate v0.2.x surface into deprecated.go
File-tree clean-up so v0.3.0 callers see a focused public API and
reviewers can spot the entire deprecation surface in one place.

- Move every v0.2.x symbol — Store, FSStore, RetrievalStore, CachedStore,
  legacy KnowledgeNode/Config/RegisterNode/Schema, NewSearchTool/NewAddTool,
  ChangeNotifier/Reloader, Document/SearchResult/SearchOptions/Chunk,
  ChunkDocument/RankResults/RRFMerge/ScoreChunk/parseFrontmatter — into
  a single sdk/knowledge/deprecated.go (~2.3k lines). Replacement index
  lives in the file's package doc.
- Rename node_v2.{go,_test.go} → node.{go,_test.go};
  tool_v2.{go,_test.go} → tool.{go,_test.go};
  reloader_v2.{go,_test.go} → reloader.{go,_test.go}.
  Symbol names keep the *Service* suffix until v0.3.0 (per arbitration
  14=B) so callers and watcher continue to compile unchanged.
- Slim types.go to the v0.3.0 model only (ContextLayer, SearchMode,
  ChunkConfig, DefaultChunkConfig, DefaultThreshold). Moved
  Document/SearchResult/SearchOptions/Chunk to deprecated.go.
- Slim hybrid.go to CosineSimilarity only; RRFMerge/rrfKey moved to
  deprecated.go.
- Slim search.go to textsearch aliases; RankResults/ScoreChunk/
  parseFrontmatter moved to deprecated.go.
- Delete chunker.go (legacy ChunkDocument moved to deprecated.go);
  chunking.go remains as the v0.3.0 chunker.
- Move ReloaderOptions onto reloader.go (shared by both EventReloader
  and the deprecated Reloader).
- Move DatasetQuery into node.go (not deprecated; reused by
  KnowledgeNodeConfig).
- Drop legacy *_test.go files (cached_store/embedder/fsstore/pipeline/
  retrieval_store/search/search_bench/hybrid/chunker/tool) per
  arbitration 13=B; v0.3.0 coverage lives in service_test, retriever_test,
  reloader_test, node_test, tool_test.

Tests: full sdk/... + sdkx/... pass.
Made-with: Cursor
