+++ sdk/knowledge/node_test.go
