+++ sdk/knowledge/reloader_test.go
