+++ sdk/knowledge/types.go
@@ -1,13 +1,28 @@
-// Package knowledge provides a side-effect-free knowledge base library:
-// document storage, chunking, tokenization, and BM25 / semantic / hybrid
-// retrieval over layered context (L0 abstract, L1 overview, L2 chunks).
+// Package knowledge implements the v0.3.0 layered knowledge base:
+// document storage, chunking, tokenization, and BM25 / vector / hybrid
+// retrieval over three context layers (L0 abstract, L1 overview,
+// L2 chunk detail).
 //
-// Storage operations (Store.AddDocument, AddDocuments) only persist raw
-// content and update search indexes; they do not call out to an LLM. To
-// derive L0/L1, use the stateless GenerateDocumentContext and
-// GenerateDatasetContext helpers and publish results back through the
-// FSStore setters and sidecar writers. This keeps scheduling, retries,
-// caching, and persistence concerns owned entirely by the caller.
+// Architecture
+//
+//   - Service       (this package): orchestrates DocumentRepo / ChunkRepo /
+//     LayerRepo, normalises Query and stamps DerivedSig
+//     so callers see a single coherent contract.
+//   - factory       (sdk/knowledge/factory): wires Service against either
+//     filesystem-backed (factory.NewLocal) or
+//     retrieval.Index-backed (factory.NewRetrieval) repositories.
+//   - SearchEngine  (this package): runs Retrievers in parallel, fuses with
+//     a Ranker (RRF by default).
+//   - EventReloader (this package): debounces ChangeEvents and triggers
+//     Service.Rebuild with the smallest possible scope.
+//
+// L0/L1 derivation (GenerateDocumentContext / GenerateDatasetContext) is
+// kept external to Service so callers own scheduling, retry and
+// persistence policy.
+//
+// Migration: every v0.2.x symbol survives in deprecated.go (tagged
+// // Deprecated:) until v0.3.0; consult deprecated.go for the full
+// new-name index.
 package knowledge
 
 // ContextLayer indicates the granularity of a search result.
@@ -23,34 +38,10 @@ const (
 	LayerDetail   ContextLayer = "L2" // full chunk content
 )
 
-// Document represents a knowledge base document.
-//
-// Deprecated: use SourceDocument (raw content + Version) and DerivedLayer
-// (L0/L1) separately. Document conflates the two and is removed in v0.3.0.
-type Document struct {
-	Name     string            `json:"name"`
-	Content  string            `json:"content"`
-	Abstract string            `json:"abstract,omitempty"` // L0
-	Overview string            `json:"overview,omitempty"` // L1
-	Metadata map[string]string `json:"metadata,omitempty"`
-}
-
-// SearchResult represents a single search hit with its relevance score.
-//
-// Deprecated: use Hit. SearchResult is removed in v0.3.0.
-type SearchResult struct {
-	Content    string         `json:"content"`
-	Score      float64        `json:"score"`
-	DocName    string         `json:"doc_name,omitempty"`
-	ChunkIndex int            `json:"chunk_index,omitempty"`
-	Layer      ContextLayer   `json:"layer"`
-	Metadata   map[string]any `json:"metadata,omitempty"`
-}
-
 // SearchMode chooses the retrieval algorithm.
 //
 // v0.3.0 final values are explicit strings; legacy callers that pass the
-// empty string are normalised to ModeBM25 by Mode.Resolve().
+// empty string are normalised to ModeBM25 by ResolveMode().
 //
 // Deprecated names:
 //   - ModeSemantic remains for backwards compatibility; new code should
@@ -65,17 +56,6 @@ const (
 	ModeHybrid   SearchMode = "hybrid"
 )
 
-// SearchOptions configures a knowledge search query.
-//
-// Deprecated: use Query. The MaxLayer→Layer rename and ScopeAllDatasets
-// fan-out live on Query. SearchOptions is removed in v0.3.0.
-type SearchOptions struct {
-	TopK      int          `json:"top_k,omitempty"`
-	MaxLayer  ContextLayer `json:"max_layer,omitempty"`
-	Threshold float64      `json:"threshold,omitempty"`
-	Mode      SearchMode   `json:"mode,omitempty"`
-}
-
 // ChunkConfig controls document chunking.
 type ChunkConfig struct {
 	ChunkSize    int `json:"chunk_size,omitempty"`
@@ -87,14 +67,6 @@ func DefaultChunkConfig() ChunkConfig {
 	return ChunkConfig{ChunkSize: 512, ChunkOverlap: 64}
 }
 
+// DefaultThreshold is the BM25-scale relevance floor used by legacy
+// search paths. Service-driven Search uses Query.Threshold instead.
 const DefaultThreshold = 0.1
-
-// Chunk represents a segment of a document.
-//
-// Deprecated: use DerivedChunk. Chunk is removed in v0.3.0.
-type Chunk struct {
-	DocName string `json:"doc_name"`
-	Index   int    `json:"index"`
-	Content string `json:"content"`
-	Offset  int    `json:"offset"`
-}
